### 2022-03-23 #4267 @editorialbot
[PRE REVIEW]: Mechkit: A continuum mechanics toolkit in Python
**Submitting author:** @JulianKarlBauer (<a href="http://orcid.org/0000-0002-4931-5869">Julian Karl Bauer</a>)
**Repository:** https://github.com/JulianKarlBauer/mechkit
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.9
