### 2022-04-01 #4267 @Kevin-Mattheus-Moerman
@nilsmeyerkit @moorepants @KeitaNakamura would you be interested in reviewing this work, on Python based continuum mechanics, for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org)?
JOSS reviews focus on [evaluating the software](https://github.com/JulianKarlBauer/mechkit) and on reviewing the [short paper](https://github.com/openjournals/joss-papers/blob/joss.04267/joss.04267/10.21105.joss.04267.pdf).

If you are interested you can comment in this issue. Thanks.
