### 2022-04-01 #4267 @Kevin-Mattheus-Moerman
> Thank you for the suggestion, but as a co-author that seems to be a conflict of interest.

@nilsmeyerkit my apologies I did not see you as a contributor to the software but see now you are on the paper. Sorry for the mix-up. FYI if you'd like to suggest reviewers at this point that would be helpful too. Thanks.
