+++ README.md
@@ -53,12 +53,15 @@ These notebooks and corresponding scripts can be found [here](docs/source/notebo
 
 ## Acknowledgment
 
-The research documented in this repository has been funded by the German Research Foundation (DFG) within the
-International Research Training Group [“Integrated engineering of continuous-discontinuous long fiber reinforced polymer structures“ (GRK 2078)][grk_website].
-The support by the [German Research Foundation (DFG)][dfg_website] is gratefully acknowledged.
+The research documented in this repository has been funded by the 
+[German Research Foundation (DFG, Deutsche Forschungsgemeinschaft)][dfg_website] - project number [255730231][dfg_project].
+The support by the German Research Foundation within the International Research Training Group 
+[“Integrated engineering of continuous-discontinuous long fiber reinforced polymer structures“ (GRK 2078)][grk_website]
+is gratefully acknowledged.
 
 [grk_website]: https://www.grk2078.kit.edu/
 [dfg_website]: https://www.dfg.de/
+[dfg_project]: https://gepris.dfg.de/gepris/projekt/255730231
 
 [url_license]: LICENSE
 [url_latest_doi]: https://doi.org/10.5281/zenodo.3898703
