### 2022-05-10 #4267 @Kevin-Mattheus-Moerman
Hi @lizarett thanks, sure you can join if interested.

We used to aim for about 2-3 weeks for reviews but since COVID we've been much more flexible. If you need more time that is fine too. Let me know if you are still interested. Would you mind sharing your full name and affiliation?
