### 2022-05-11 #4267 @Kevin-Mattheus-Moerman
@lizarett sure, here is an example review: https://github.com/openjournals/joss-reviews/issues/1441. As you can see, each reviewer gets a set of check boxes to guide you through the review process. The review focuses on the software and the short paper as mentioned :point_up: above.

You can email me at kevin.moerman@nuigalway.ie.

Let me know if you have any questions and if you are interested in helping us. Thanks!
