### 2022-05-12 cbbffbe
Add missing DOIs following editorialbot
@editorialbot in 
https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1124216841_
