+++ paper/paper.bib
@@ -146,7 +146,8 @@ journal      = {International Journal of Engineering Science},
 publisher    = {Elsevier},
 volume       = 8,
 number       = 6,
-pages        = {475--481}
+pages        = {475--481},
+doi          = {https://doi.org/10.1016/0020-7225(70)90024-8}
 }
 
 
@@ -158,7 +159,8 @@ pages        = {475--481}
 	publisher    = {The Society of Rheology},
 	volume       = 31,
 	number       = 8,
-	pages        = {751--784}
+	pages        = {751--784},
+    doi          = {10.1122/1.549945}
 }
 
 @software{fiberoripy,
