### 2022-05-12 #4389 @JulianKarlBauer
> @JulianKarlBauer can you check those potentially missing DOI's point_up ?

Thank you for pointing this out! I added the missing DOIs and regenerated the pdf.
