### 2022-05-12 #4389 @Kevin-Mattheus-Moerman
@JulianKarlBauer nearly there, not the invalid one :point_up:, fyi you can also run `@editorialbot check references`.
