### 2022-05-12 781da1f
Fix DOI, remove prefix
@editorialbot in 
https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1125094378_
