+++ paper/paper.bib
@@ -147,7 +147,7 @@ publisher    = {Elsevier},
 volume       = 8,
 number       = 6,
 pages        = {475--481},
-doi          = {https://doi.org/10.1016/0020-7225(70)90024-8}
+doi          = {10.1016/0020-7225(70)90024-8}
 }
 
 
