### 2022-05-23 #4389 @nicoguaro
I have done a first pass on the package. I have some comments:

- I don't see a clear statement of need on the README or docs.
- There isn't an API section on the docs. My guess is that it is already there in the different sections that appear on ReadTheDocs.
- There isn't a section or file specifying how to contribute to the project.

I will resume the review after these comments have been addressed.
