+++ docs/index.rst
@@ -1,16 +1,6 @@
 mechkit
 =======
 
-Statement of need
------------------
-
-The mathematical operators for material modeling and linear elasticity can be expressed in a compact manner. However, their representation is not unique and the relation between different representations is nontrivial. To the best knowledge of the authors, there is no common library for the mathematical operators to this day. This presents an obstacle with regard to the exchange and the reliability of research code and leads to negative consequences for the interpretation and comparison of results.
-
-The main goal of mechkit is to provide reusable research code that increases the reliability of results. It is intended to accelerate and simplify further research.
-
-
-----------------------------------------------
-
 Examples
 --------
 
@@ -38,6 +28,17 @@ API
     /source/visualization
 
 ----------------------------------------------
+
+
+Statement of need
+-----------------
+
+.. toctree::
+   /source/statement_of_need
+
+
+----------------------------------------------
+
 
 References
 ----------
