+++ docs/index.rst
@@ -1,3 +1,4 @@
+=======
 mechkit
 =======
 
