### 2022-05-25 #4389 @JulianKarlBauer
Hi @nicoguaro
thank you very much for starting your review!
Please let me address the points of concern.

-----------------------------------

>
>     * There isn't a section or file specifying how to contribute to the project.
>
Thank you, indeed a file `CONTRIBUTING.md` only had been added to the `paper`-branch, which had been used to create the paper draft and prepare the review. I merged the `paper`-branch into master and will keep both identical side by side. As I specified the `paper`-branch during submission to JOSS, I'll keep it alive till the end of the review.

-----------------------------------
>
>     * I don't see a clear statement of need on the README or docs.
>
This is true, I copied the "Statement of need" from the paper to the landing page of the docs. Personally I think that the README.md should not contain a statement of need, this is why I added it to the docs. Is this ok for you?
(see b8052796ed936d32b8ace1160bea27f5cc596458)

-----------------------------------

>
>     * There isn't an API section on the docs. My guess is that it is already there in the different sections that appear on ReadTheDocs.
>
I added headings to the different parts of the documentation now looking like:
![image](https://user-images.githubusercontent.com/55383008/170219852-5785f688-2c19-4427-8673-89023bf0702b.png)

Is this ok from your point of view?

-----------------------------------

Thank you again for the review and the valuable feedback!
