### 2022-06-07 #4389 @nicoguaro
> Is this ok from your point of view?

It is better. Although, I would add an API section in the left menu bar to group all the different modules.

Regarding the paper, I don't see any mention of similar packages. I do not know if there are any besides [SymPy's module](https://docs.sympy.org/latest/modules/physics/continuum_mechanics/index.html) (not so similar) and [continuum_mechanics](https://continuum-mechanics.readthedocs.io/en/latest/?badge=latest) (disclaimer, I am the author).
