+++ docs/index.rst
@@ -3,11 +3,37 @@ mechkit
 =======
 
 .. toctree::
-    :maxdepth: 1
+    :maxdepth: 4
 
    /source/statement_of_need
-   /source/example_notebooks
-   /source/api
+
+----------------------------------------------
+
+Examples
+--------
+
+.. toctree::
+    :maxdepth: 1
+
+    /source/example_notebooks
+
+----------------------------------------------
+
+API
+---
+
+.. automodule:: mechkit
+
+.. toctree::
+    :maxdepth: 4
+
+    /source/fabric_tensors
+    /source/notation
+    /source/material_index
+    /source/tensors
+    /source/utils
+    /source/operators
+    /source/visualization
 
 
 ----------------------------------------------
