+++ docs/source/api.rst
@@ -1,15 +1,0 @@
-API
----
-
-.. automodule:: mechkit
-
-.. toctree::
-    :maxdepth: 4
-
-    /source/fabric_tensors
-    /source/notation
-    /source/material_index
-    /source/tensors
-    /source/utils
-    /source/operators
-    /source/visualization
