### 2022-06-07 #4389 @JulianKarlBauer
> > Is this ok from your point of view?
>
> It is better. Although, I would add an API section in the left menu bar to group all the different modules.
>

Thanks, I'll fight with Sphinx to clean the lean toctree.

> Regarding the paper, I don't see any mention of similar packages. I do not know if there are any besides [SymPy's module](https://docs.sympy.org/latest/modules/physics/continuum_mechanics/index.html) (not so similar) and [continuum_mechanics](https://continuum-mechanics.readthedocs.io/en/latest/?badge=latest) (disclaimer, I am the author).

Thank you for mentioning these packages which also refer to continuum mechanics. However, the tasks accomplished by these packages are quite varying. This is probably caused by the fact that the phrase continuum mechanics describes a high-level topic. As far as I know, no packages similar to mechkit exist. This is the reason, why I did not list others.

SymPy's module implements beam-theory and therefore should refer to "structural mechanics" instead of continuum mechanics. Continuum_mechanics is a great package for a rather specific selection of continuum mechanics topics, like mechkit. But as far as I see, no common tasks are solved.
