### 2022-06-21 #4389 @nicoguaro
> @nicoguaro has @JulianKarlBauer address your comments to your satisfaction? ☝️

@Kevin-Mattheus-Moerman, partially, I would argue that it should be clearer what is the difference with other packages. At least, since the name includes a broad topic _Mechkit: A continuum mechanics toolkit in Python_. I am still pending on checking the code and the tests.
