### 2022-06-22 JulianKarlBauer/mechkit#46 @nicoguaro
pytest not installed following README instructions
When you run

    python setup.py develop

Pytest is not installed and you can't run the test suite.
