### 2022-06-25 JulianKarlBauer/mechkit#46 @JulianKarlBauer
Hi nicoguaro,
thank you for pointing this out. So far, this is intended behavior.
Why do you expect that installing in "develop"-mode would install pytest?

If I do `python setup.py develop` on the source code of [continuum_mechanics](https://github.com/nicoguaro/continuum_mechanics), test requirements are also not met, which from my personal point of view is fine.

If you point me towards an example where installing in develop mode is used to install test requirements, I rethink whether this is to be considered as a feature.
