+++ .github/workflows/python_test.yml
@@ -19,8 +19,7 @@ jobs:
     - name: Install Package
       run: |
         python -m pip install --upgrade pip
-        pip install pytest
-        python setup.py install
+        python setup.py develop
     - name: Test with pytest against installed package
       run: |
         pytest
