### 2022-06-25 JulianKarlBauer/mechkit#46 @JulianKarlBauer
Another thought: The tests are run by Github-CI before pushing to PyPI. I don't think any user will run the tests. If so, instructions on running the tests are found in .github/workflow/... files. I think this procedure is quite common.
