### 2022-06-25 #4389 @JulianKarlBauer
Thank you @nicoguaro for pointing out, that the comparison with other packages should be extended.

After having looked at all 23 repositories tagged with "continuum-mechanics" on Github, it becomes clear, that our repositories are quite isolated when it comes to Python packages for continuum-mechanics. Most of the remaining 21 repositories are even more specific and only a few are written in Python, as Fortran is standard in the finit-element-method context.

If you don't mind, I'll focus on comparison to Sympy and your package and propose adding the following to the paper draft:

"
The number of Python packages in the area of continuums mechanic is quite small. In addition to the symbolic treatment of structural mechanics problems with Sympy, the Python package continuum_mechanics supports visualization of second-order tensors as well as compact representation of coordinate transformations and tensor analysis operators. Compared to these packages, mechkit focuses on notation, material containers and algebraic operators, and provides these methods as a toolkit for dependent Python packages such as mechmean.
"

If this addition with suitable citation is ok for you, I'll propose it to my co-authors and change the pdf.
