### 2022-06-29 JulianKarlBauer/mechkit#46 @nicoguaro
I am not thinking as a user in this context but as a contributor. In that case, it is common practice to run tests locally before doing a PR. I would guess that's the public intended for that [information](https://github.com/JulianKarlBauer/mechkit/tree/master/test).
