### 2022-06-29 f4a0a1c
Revert "Use continuum_mechanics as template, without success"
This reverts commit eb223d29088b07f63b5b3e8b60cf64cd87160b15.
