+++ .github/workflows/python_test.yml
@@ -19,7 +19,8 @@ jobs:
     - name: Install Package
       run: |
         python -m pip install --upgrade pip
-        python setup.py develop
+        pip install pytest
+        python setup.py install
     - name: Test with pytest against installed package
       run: |
         pytest
