### 2022-06-29 b938d43
Add step: Install pytest
