+++ test/README.md
@@ -1,14 +1,14 @@
-# Run tests against installation (recommended)
+## Run tests against installation (recommended)
 - Install mechkit in development mode `python setup.py develop`
 - Install pytest `pip install pytest`
 - Run tests `pytest`
 - Uninstall `python setup.py develop --uninstall`
 
-# Run tests against local files
+## Run tests against local files
 - Install mechkit and pytest (see steps above)
 - Run tests with `python -m pytest`  
 
-# Run [specific](https://stackoverflow.com/a/62804929/8935243) tests
+## Run [specific](https://stackoverflow.com/a/62804929/8935243) tests
 
 Example from within ipython
 ```python
