### 2022-06-29 JulianKarlBauer/mechkit#46 @JulianKarlBauer
Thank you so much for reminding me on the contributor instructions on tests. Now I realize the problem. I added a note on installing pytest before running the tests. Are you fine with this solution? Please see #47 for a proposal.
