### 2022-06-29 #4389 @nicoguaro
@Kevin-Mattheus-Moerman, I have reviewed the software taking into account the end-user and contributor perspectives. Following I mention some points regarding this submission:

- The software is useful for researchers in material modeling; I would have used a software like this during my Ph.D., for example.
- The software is related to continuum mechanics but, right now, it seems to be more focused on constitutive tensors. This fit perfectly within *material modeling*.
- The visualization module of the package does not visualize anything right now. It does give "Young moduli" in different directions and one could construct a surface with that information as in *Graphical representation of the generalized Hooke’s law*.
- I created an [issue](https://github.com/JulianKarlBauer/mechkit/issues/46) regarding a dependency missing for testing during development.

Regarding the lack of reviews (besides mine) should we ask on the Computational Mechanics community on Twitter?
