### 2022-07-03 #4389 @likask
@JulianKarlBauer My initial impression is that it is valuable code. I might use it in my work or recommend it to my colleagues.

- To "Contributing guide" add description about typical contribution scheme, fork->modify->test->pull request.

- I would advise mentioning in the paper the difference between this package and other tensor algebra codes. Explicitly mention some packages (e.g. https://en.wikipedia.org/wiki/Tensor_software). Studying this subject, I see that mechkit offers tensor algebra functionality and tools specifically for analysis, construction, and managing description in various notations used in solid continuum mechanics. That not necessary would be clear for someone looking for a tool like this one.

More generic comments for the future:

- Interesting development would be a visualisation of tensors using packages such as PyVista, and creating a toolset for this.

- It would be interesting to see the extension for non-classical continuum problems, such as second-order continuum, micromorphic, etc., where third-order tensors are used.
