### 2022-07-05 #4389 @lizarett
The code appears to offer useful functionality of conversion from one notation to the other, examples help understanding the use cases. A possible suggestion would be to have in the description a sort of summary table or a list of all new function names with a list of all corresponding input variables. This may be just one more way to learn the package faster.
