### 2022-07-12 JulianKarlBauer/mechkit#46 @nicoguaro
> Thank you so much for reminding me on the contributor instructions on tests. Now I realize the problem. I added a note on installing pytest before running the tests. Are you fine with this solution? Please see #47 for a proposal.

It looks OK to me.
