### 2022-08-12 JulianKarlBauer/mechkit#48 @JulianKarlBauer
Improve contributing instructions
Addressing this [comment](https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1173060634), the contributing instructions should be improved and switch from "Github flow" to "Fork and Pull request".
