### 2022-08-12 630cada
Change case titles and simplify
