+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
-# Contributing guide of mechkit
+# Contributing Guide of Mechkit
 
 Thank you for investing your time in contributing to this project!
 
@@ -6,7 +6,7 @@ Please follow the [Code of Conduct](./CODE_OF_CONDUCT.md) to keep the community 
 
 In this guide you will get an overview of the contribution workflow from opening an issue, creating a PR, reviewing, and merging the PR.
 
-## New contributor guide
+## New Contributor Guide
 
 To get an overview of the project, read the [README](README.md).
 Github offers the following some resources to help you get started with open source contributions:
@@ -16,13 +16,13 @@ Github offers the following some resources to help you get started with open sou
 - [Collaborating with pull requests](https://docs.github.com/en/github/collaborating-with-pull-requests)
 
 
-## Workflow: Issues / Problems / Bugs
+## Bugs / Issues
 
 We don't like bugs, and are therefore grateful for any possibility to fix them.
 Please report any problems by opening an issue.
 Screenshots as well as minimal working examples are highly appreciated, as they ease the understanding of the problem.
 
-## Workflow: Code Contributions
+## Workflow: Fork -> Modify -> Test -> Pull Request
 
 The main goal of the project is the reuse of research code to increase the reliability of research results and to accelerate and simplify further research.
 This objective implies the need for continuous extension and improvement. As a consequence, code contributions of any kind are welcome.
