### 2022-08-12 JulianKarlBauer/mechkit#51 @JulianKarlBauer
In paper: Compare with other tensor packages
Addressing [this review comment](https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1173060634), a comparison with multi-purpose tensor packages should be contained within the mechki-JOSS-paper.
