### 2022-08-12 3eb860a
Add reference to general-purpose tensor packages
