+++ paper/paper.bib
@@ -215,3 +215,19 @@ license = {GPL-3.0},
 year         = 2022,
 }
 
+@article{tensorly,
+  title={Tensorly: Tensor learning in python},
+  author={Kossaifi, Jean and Panagakis, Yannis and Anandkumar, Anima and Pantic, Maja},
+  journal={arXiv preprint arXiv:1610.09555},
+  year={2016}
+}
+
+@article{tenpy,
+  title={Efficient numerical simulations with tensor networks: Tensor Network Python (TeNPy)},
+  author={Hauschild, Johannes and Pollmann, Frank},
+  journal={SciPost Physics Lecture Notes},
+  pages={005},
+  year={2018}
+}
+
+
