### 2022-08-12 #4389 @JulianKarlBauer
> @JulianKarlBauer My initial impression is that it is valuable code. I might use it in my work or recommend it to my colleagues. The key is that transformations related to various notations used in continuum mechanics are tested, and implementing them from scratch in generic tensor code consumes time for validation.
> ...

Thank you @likask for the valuable feedback and sorry for the late reply, I survived Covid and earned a phd in the meantime :)

> More generic comments for the future:
>
>     * Interesting development would be a visualisation of tensors using packages such as PyVista, and creating a toolset for this.
>
>     * It would be interesting to see the extension for non-classical continuum problems, such as second-order continuum, micromorphic, etc., where third-order tensors are used.
Adding routines and methods for non-classical continuum problems is indeed a valuable future topic! Are you an expert in this field?

--------------------------------------------

>
>     * To "Contributing guide" add description about typical contribution scheme, fork->modify->test->pull request.
>
Thank you for pointing out the possible improvement in `CONTRIBUTING.md`. I followed your advice to specify a "Fork, Modify, Test, Pull request"-workflow explicitly. Are you happy with the changes made by this [pull request](https://github.com/JulianKarlBauer/mechkit/pull/49)? If you'd like to see more improvements, please feel free to reopen [issue 48](https://github.com/JulianKarlBauer/mechkit/issues/48).

>
>     * I would advise mentioning in the paper the difference between this package and other tensor algebra codes. Explicitly mention some packages (e.g. https://en.wikipedia.org/wiki/Tensor_software). Studying this subject, I see that mechkit offers tensor algebra functionality and tools specifically for analysis, construction, and managing description in various notations used in solid continuum mechanics. That not necessary would be clear for someone looking for a tool like this one.
>
Thank you motivating a comparison with other tensor packages in the paper.
I opened this [issue](https://github.com/JulianKarlBauer/mechkit/issues/51) and added a references and highlighted differences to other multi-purpose tensor packages in Python in this [pull request](https://github.com/JulianKarlBauer/mechkit/pull/52). Does this fit your intention?
