### 2022-08-12 #4389 @JulianKarlBauer
> The code appears to offer useful functionality of conversion from one notation to the other, examples help understanding the use cases. A possible suggestion would be to have in the description a sort of summary table or a list of all new function names with a list of all corresponding input variables. This may be just one more way to learn the package faster.

Thank you @lizarett for pointing out the potential to improve the documentation, I guess I have a blind spot here.
Up to now my strategy has been to give reasonable [examples](https://mechkit.readthedocs.io/en/latest/source/example_notebooks.html), as examples are straight forward. Are you more interested in a kind of list of functions like in the sub-chapters of the docs or details on argument types for some of the functions?
