### 2022-08-12 #4389 @lizarett
@JulianKarlBauer i mean list of functions collected in one place. in my view, table formats with functions-explanation-arguments work well as a part of the documentation. please, see examples here down the page https://uk.mathworks.com/help/matlab/math/choose-an-ode-solver.html , maybe, something like this is suitable.
