### 2022-08-22 JulianKarlBauer/mechkit#53 @JulianKarlBauer
In paper: Add overview tables
Add overview tables following suggestion by [reviewer](https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1174758863).
