+++ docs/source/operators.rst
@@ -1,6 +1,16 @@
 operators
 ---------
 
++-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
+| Operator                                                    | Example configuration                                              | Vectorized  | Supported tensor orders  |
++=============================================================+====================================================================+=============+==========================+
+| mechkit.operators.Sym(axes).__call__                        | first two axes: axes=[0,1]                                         | no          | any                      |
++-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
+| mechkit.operators.Sym_Fourth_Order_Special(label).__call__  | supported labels: ["left", "right", "major", "inner", "complete"]  | no          | 4                        |
++-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
+| mechkit.operators.dev                                       | none                                                               | no          | 2, 4                     |
++-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
+
 .. automodule:: mechkit.operators
     :members:
     :undoc-members:
