+++ docs/source/operators.rst
@@ -1,15 +1,19 @@
 operators
 ---------
 
-+-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
-| Operator                                                    | Example configuration                                              | Vectorized  | Supported tensor orders  |
-+=============================================================+====================================================================+=============+==========================+
-| mechkit.operators.Sym(axes).__call__                        | first two axes: axes=[0,1]                                         | no          | any                      |
-+-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
-| mechkit.operators.Sym_Fourth_Order_Special(label).__call__  | supported labels: ["left", "right", "major", "inner", "complete"]  | no          | 4                        |
-+-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
-| mechkit.operators.dev                                       | none                                                               | no          | 2, 4                     |
-+-------------------------------------------------------------+--------------------------------------------------------------------+-------------+--------------------------+
+
++-------------------------------------------------------------+--------------------------+-------------+---------------------------------------------------------------------------------------------------------------------+
+| Operator                                                    | Supported tensor orders  | Vectorized  | Example configuration                                                                                               |
++=============================================================+==========================+=============+=====================================================================================================================+
+| mechkit.operators.Sym(axes).__call__                        | any                      | no          | axes=[0,1] symmetrizes the first two axes                                                                           |
++-------------------------------------------------------------+--------------------------+-------------+---------------------------------------------------------------------------------------------------------------------+
+| mechkit.operators.Sym_Fourth_Order_Special(label).__call__  | 4                        | no          | label="major" takes the major symmetric part, supported labels are ["left", "right", "major", "inner", "complete"]  |
++-------------------------------------------------------------+--------------------------+-------------+---------------------------------------------------------------------------------------------------------------------+
+| mechkit.operators.dev                                       | 2, 4                     | no          | none                                                                                                                |
++-------------------------------------------------------------+--------------------------+-------------+---------------------------------------------------------------------------------------------------------------------+
+
+
+--------------------------------------------------------
 
 .. automodule:: mechkit.operators
     :members:
