### 2022-08-22 #4389 @lizarett
> > @JulianKarlBauer i mean list of functions collected in one place. in my view, table formats with functions-explanation-arguments work well as a part of the documentation. please, see examples here down the page https://uk.mathworks.com/help/matlab/math/choose-an-ode-solver.html , maybe, something like this is suitable.
>
> @lizarett Could you have a look at the overview tables for the modules [material](https://mechkit.readthedocs.io/en/latest/source/material.html), [notation](https://mechkit.readthedocs.io/en/latest/source/notation.html) and [operators](https://mechkit.readthedocs.io/en/latest/source/operators.html) I added in [issue](https://github.com/JulianKarlBauer/mechkit/issues/53)?

thank you, looks good to me
