### 2022-09-11 #4389 @Kevin-Mattheus-Moerman
@nicoguaro, @likask, @lizarett can you provide an update on review progress or what is still outstanding? It looks like @JulianKarlBauer has addressed several points.

@lizarett can you check if you are now happy to tick that last box and recommend acceptance?

@likask can you check the response to your comments? :point_up: https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1212835958

@nicoguaro you have some boxes unticked, can you have a look if there are any updates, or provide points that still need to be addressed?

Thanks all for your continued help!
