### 2022-09-13 #4389 @JulianKarlBauer
> > @nicoguaro you have some boxes unticked, can you have a look if there are any updates, or provide points that still need to be addressed?
>
> All of them have been addressed except for the "state of the field". @JulianKarlBauer mentioned something before about it, but I don't see anything about it in the compiled version of the paper.

Sorry, I had not yet merged [pull request](https://github.com/JulianKarlBauer/mechkit/pull/52) mentioned above. After having merged it, are your concerns addressed?
