### 2022-09-13 #4389 @nicoguaro
> Sorry, I had not yet merged [pull request](https://github.com/JulianKarlBauer/mechkit/pull/52) mentioned above. After having merged it, are your concerns addressed @nicoguaro ?

Yes.

@Kevin-Mattheus-Moerman, I have marked all the boxes now and recommend the publication.
