### 2022-09-16 #4389 @likask
@JulianKarlBauer @Kevin-Mattheus-Moerman

Hi, I am happy from your modifications. Green light from me.

Yes, some time ago I was working on second-order continuum description, dealing with third order tensors was painful in vector notation.
