### 2022-09-16 #4389 @Kevin-Mattheus-Moerman
Great @likask can  you chech those last boxes? :point_right: https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1173047773
