### 2022-10-07 #4389 @JulianKarlBauer
> @JulianKarlBauer can you check that potentially missing DOI? point_up

@Kevin-Mattheus-Moerman Thank you! Sorry I could have checked the references on my own...
Repeated the checks and build.
