### 2022-10-08 #4389 @Kevin-Mattheus-Moerman
@JulianKarlBauer I read your paper and have two remaining very minor points:

- [ ] Please add city names to all affiliations
- [ ] For `several ... , exist but mainly...` it looks like the comma should go after `exist` (but feel free to disagree)
