### 2022-10-09 975dbf0
Add city names in affiliation
https://github.com/openjournals/joss-reviews/issues/4389#issuecomment-1272254270
