### 2022-10-09 #4389 @JulianKarlBauer
* [x]  Please add city names to all affiliations
* [x]  For `several ... , exist but mainly...` it looks like the comma should go after `exist` (but feel free to disagree)

@Kevin-Mattheus-Moerman Thank you! Fixed both.
