### 2022-10-10 #4389 @Kevin-Mattheus-Moerman
@JulianKarlBauer Great, looks like we are good to proceed. At this point can you please:

- [ ] Archive a copy of the reviewed work on [ZENODO](https://zenodo.org/). We recommend that you do so for a newly minted/tagged release, however this is up to you. Some find these automated steps useful: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
- [ ] Once the archive is created please check if the title and the author names match those of the paper. You may need to manually edit your archive entry to change the title/author entries. You can also check if you can have it refer to the license properly (or leave it as Other(Open) since the archived content contains the license too). Next you may wish to add author ORCID information too. The most important things though at the title and author list.
- [ ] Once the above is complete please report back here with the archive DOI and which version tag we should use, i.e. are we still at v0.3.9 or have you created a new release.

Thanks.
