+++ docs/source/notation.rst
@@ -14,7 +14,6 @@ notation
 .. autoclass:: mechkit.notation.Converter
     :members:
     :undoc-members:
-    :noindex:
     :exclude-members: to_mandel6, to_mandel9, to_like, to_tensor
 
     .. :show-inheritance:
@@ -30,7 +29,6 @@ notation
 .. autoclass:: mechkit.notation.ExplicitConverter
     :members:
     :undoc-members:
-    :noindex:
     :exclude-members: convert
 
 
