### 2022-10-11 #4389 @JulianKarlBauer
@Kevin-Mattheus-Moerman
Perfect, thank you for the detailed instructions.
I created release [v0.4.0](https://zenodo.org/record/7185691) with DOI https://zenodo.org/record/7185691
and manually edited the title, author, license and added a funding note.
