### 2022-10-12 #4389 @JulianKarlBauer
(: thanks for the great experience of publishing in JOSS.
Thank you  @Kevin-Mattheus-Moerman for your efforts in moderation!
Thanks to the reviewers @nicoguaro @likask @lizarett for your valuable suggestions and dedication and thank you @kthyng for finalizing the process.
